backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
aiohttp
psycopg2-binary